Move Components to their own directory
